proyecto_institutos/requirements.txt
django>=4.2
djangorestframework>=3.14
pandas>=2.0
drf-yasg>=1.21
Pillow>=10.0
python-decouple>=3.8
